Add lock check to delete deposit API call and log patch user

+++ src/DigitalPreservation/Preservation.API/Features/Deposits/Requests/DeleteDeposit.cs
@@ -1,5 +1,7 @@
-﻿using DigitalPreservation.Common.Model;
+﻿using System.Security.Claims;
+using DigitalPreservation.Common.Model;
 using DigitalPreservation.Common.Model.Results;
+using DigitalPreservation.Core.Auth;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Preservation.API.Data;
@@ -7,9 +9,10 @@ using Storage.Repository.Common;
 
 namespace Preservation.API.Features.Deposits.Requests;
 
-public class DeleteDeposit(string id) : IRequest<Result>
+public class DeleteDeposit(string id, ClaimsPrincipal user) : IRequest<Result>
 {
     public string Id { get; } = id;
+    public readonly ClaimsPrincipal User = user;
 }
 
 public class DeleteDepositHandler(
@@ -21,9 +24,14 @@ public class DeleteDepositHandler(
     {
         try
         {
+            var callerIdentity = request.User.GetCallerIdentity();
             var entity = await dbContext.Deposits.SingleOrDefaultAsync(d => d.MintedId == request.Id, cancellationToken);
             if (entity != null)
             {
+                if (entity.LockedBy is not null && entity.LockedBy != callerIdentity)
+                {
+                    return Result.Fail(ErrorCodes.Conflict, "Deposit is locked by " + entity.LockedBy);
+                }
                 var storageLocation = entity.Files;
                 dbContext.Deposits.Remove(entity);
                 await dbContext.SaveChangesAsync(cancellationToken);
